+++ DataAccsess/DataAccess.csproj
@@ -8,7 +8,6 @@
 
   <ItemGroup>
     <PackageReference Include="Dapper" Version="2.1.35" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.10" />
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
     <PackageReference Include="MySqlConnector" Version="2.3.7" />
   </ItemGroup>
