--- DomainLogic/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace Logic
-{
-    public class Class1
-    {
-
-    }
-}
