+++ Gruppe10KVprototype/Views/Innmelder/InnmelderSkjema/InnmelderSkjemaResultat.cshtml
@@ -1,7 +1,7 @@
-@model Models.Entities.InnmelderIncidentFormModel
+@model Models.Entities.IncidentFormModel
 
 @{
-    ViewData["Title"] = "Form Result";
+   
     Layout = "_Layout";
 }
