+++ Gruppe10KVprototype/Views/Innmelder/InnmelderSkjema/InnmelderSkjemaView.cshtml
@@ -1,7 +1,7 @@
-@model Models.Entities.InnmelderIncidentFormModel
+@model Models.Entities.IncidentFormModel
 
 @{
-    ViewData["Title"] = "Meld inn hendelse";
+   
     Layout = "_Layout";
 }
