+++ Gruppe10KVprototype/Views/Saksbehandler/BehandleInnmeldingSaksB/BehandleInnmeldingSaksB.cshtml
@@ -1,0 +1,130 @@
+﻿@model ViewModels.BehandleInnmeldingSaksBViewModel
+
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Rapport</title>
+</head>
+<body>
+    <h1>Lene Olsen</h1>
+    <h1>- har drept skjelettet</h1>
+    <h1>nb! databasen har bare 9 oppføringer</h1>
+
+    @if (Model != null)
+    {
+        <form asp-action="BehandleInnmeldingSaksB" method="post">
+            <div>
+                <label for="Innmelding_InnmeldID">Innmeldings ID:</label>
+                <input type="number" asp-for="InnmeldingE.InnmeldID" required />
+                <button type="submit">Hent Innmelding</button>
+            </div>
+
+            @if (Model.InnmeldingE != null)
+            {
+                @if (Model.InnmeldingE.InnmeldID > 0)
+                {
+                    <div>
+                        <h2>Innmelding Detaljer:</h2>
+                        <p>InnmeldID: @Model.InnmeldingE.InnmeldID</p>
+                        <p>InnmelderID: @Model.InnmeldingE.InnmelderID</p>
+                        <p>Innmeldingstidspunkt: @Model.InnmeldingE.Innmeldingstidspunkt</p>
+                        <p>Nåværende Status: @Model.InnmeldingE.StatusID</p>
+                        <p>SaksbehandlerID: @(Model.InnmeldingE.SaksbehandlerID ?? 0)</p>
+                        <p>Siste Endring: @Model.InnmeldingE.SisteEndring</p>
+                        <p>KommuneNR: @Model.InnmeldingE.KommuneNR</p>
+                        <p>Prioritet: @Model.InnmeldingE.PrioritetID</p>
+                        <p>KartID: @Model.InnmeldingE.KartID</p>
+
+                        <div>
+                            <form asp-action="UpdateStatusAsync" method="post">
+                                <input type="hidden" asp-for="InnmeldingE.InnmeldID" />
+                                <label for="status">Velg status:</label>
+                                <select asp-for="InnmeldingE.StatusID" asp-items="Model.StatusList" class="form-control"></select>
+                                <button type="submit" class="btn btn-primary">Oppdater status</button>
+                            </form>
+
+
+                        </div>
+                    </div>
+                }
+                else
+                {
+                    <p>Ingen innmelding funnet med den angitte ID.</p>
+                }
+            }
+        </form>
+    }
+    else
+    {
+        <p>Modellen er null. Vennligst kontakt systemadministrator.</p>
+    }
+</body>
+</html>
+
+
+//
+  //  Layout = "_Layout";
+//}
+
+<!--
+
+<!DOCTYPE html>
+
+<html>
+<head>
+    <title>title</title>
+</head>
+<body>
+<div class="topplinje">
+    <div>Navn Navnesen</div>
+    <div>
+        <div>
+            <ul class="topplinjeinfo">
+                <li>Status</li>
+                <li>Innmelding ID</li>
+            </ul>
+        </div>
+    </div>
+</div>
+<div class="subjektlinje">
+    <div>Subjekt</div>
+    <div>
+        <ul class="subjektlinjeinfo">
+            <li>Kommune</li>
+            <li>Tag</li>
+        </ul>
+    </div>
+</div>
+<div class="forstebokslinje">
+    <div class="beskrivelseboks">
+        <div>Beskrivelse</div>
+        <div class="descboks">Description</div>
+    </div>
+    <div class="kartvisningboks">
+        <div>Kartvisning</div>
+        <div class="mapboks">Map View</div>
+    </div>
+</div>
+<div class="bunnlinje">
+    <div>Opplysninger om innsender</div>
+    <div class="bokslinje">
+        <div class="opplysboks">Henter noe informasjon om innsenderen</div>
+        <div class="buttonboks">
+            <div class="buttonboksEn">
+                <button>SVAR</button>
+                <button>FLAGG</button>
+            </div>
+            <div class="buttonboksTo">
+                <button>OPPDATER</button>
+                <button>DELEGER</button>
+            </div>
+            <div class="buttonboksTre">
+                <button>BEHANDLE</button>
+                <button>AVVIS</button>
+            </div>
+        </div>
+    </div>
+</div>
+</body>
+</html>
+-->
