+++ Gruppe10KVprototype/Views/Saksbehandler/KartvisningInnmeldingSaksB/KartvisningInnmeldingSaksB.cshtml
@@ -1,9 +1,9 @@
-﻿@model Models.Entities.SaksbehandlerIncidentFormModel
+﻿@model Models.Entities.IncidentFormModel
 
 @{
     ViewData["Title"] = "Innmelding Detaljer";
 }
-<link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css"/>
+<link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" />
 <h2>@Model.Subject</h2>
 
 <div id="map" style="height: 500px;"></div>
