+++ Interfaces/IInnmeldingERepository.cs
@@ -1,16 +1,11 @@
-﻿
-using Models.Entities;
+﻿using Models.Entities;
 
-
-namespace Interfaces
+namespace Interface
 {
-    public interface ISaksbehandlerRepository
-    {
-        // Metoder for incident_form tabellen
-        Task<IEnumerable<SaksbehandlerIncidentFormModel>> GetAllAdviserFormsAsync();
-        Task<SaksbehandlerIncidentFormModel> GetAdviserFormByIdAsync(int id);
 
-        // Metoder for INNMELDING tabellen
+
+    public interface IInnmeldingERepository
+    {
         Task<IEnumerable<InnmeldingEModel>> GetAllInnmeldingerAsync();
         Task<InnmeldingEModel> GetInnmeldingByIdAsync(int innmeldID);
         Task<StatusEnum> GetStatusByInnmeldIdAsync(int innmeldID);
