+++ ViewModels/ViewModels.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\DomainLogic\Logic.csproj" />
+   
     <ProjectReference Include="..\Models\Models.csproj" />
   </ItemGroup>
